Add encode-decode roundtrip test with generated data for all packets types.
Round-tripping doesn't garantee that we follow the spec, but at least garantees that we're
self-consistent, and indeed exposes a bug in the current publish implementation.

The tests are implemented using proptest, giving greater confidence that all corner-cases were
tested. However, it raises the minimum supported rust version from 1.31 to 1.32, (for test builds
only).