Refactor PacketIdentifier from u16 to NonZeroU16.
This matches the MQTT spec which mandates non-zero pids. Mqttrs will now error out when trying to
encode/decode a PacketIdentifier of 0, like for example mosquitto does. It also reduces the memory
usage of `Option<PacketIdentifier>`.

Fixes #14

Using PacketIdentifier is now more verbose, but mqttrs user won't be affected as much as mqttrs
itself. While we're breaking the API, it's tempting to shorten `PacketIdentifier` to `Pid`, as we've
already got `QoS` and will hopefully soone get `QoSPid`.

PacketIdentifyer::new() is pretty much a TryInto trait, but using that would have raised the minimum
rust version requirement for little convenience.