Refactor: Tweak API visibility.
Hopefully I got the balance right, but we can always reexport missing APIs if this cull was too
eager.

One note about PacketType: it got moved to `packet.rs` and can now be obtained from a `Packet`
enum. Most end-users will just `macth` on `Packet` variants, but it can be interesting to match a
single type or to gather stats about all types.