Add `impl From<Foo> for Packet` for all packet subtypes.
It makes instanciating a `Packet` nicer/shorter to write, but the effect is limited because
`encode()` still takes an `&Packet`, so the caller needs to call `.into()` itself. Changing
`encode()` to consume the packet would solve that and would enable us to later have a zerocopy
implementation, but that seems a step too far for this iteration, and it may be handy to be able to
hold on to the original `Packet` after encoding it, for QoS implementaion purposes.