Upgrade bytes dep to 0.5
This is a breaking chang due to bytes's changed API and MSRV increasing from 1.32 to 1.39. Bytes 0.5
has a cleaner API, grows the buffer as needed, supports no_std, and a few other niceties. It is
required by newer versions of tokio and probably other important crates.

While at it, did some refactoring like s/header_u8/header/ (the new bytes API makes that clear
enough), s/buffer/buf/ (we had a mix), reducing copy-paste in encode_test.rs, and a couple of other
readability changes.

The upgrade is a good thing but it might be problematic for some users (as it is tied to async
tokio), so we should find ways to reduce bytes's API exposure and/or reexport bytes to facilitate
using multiple versions.