Partial switch from BytesMut to Buf/BufMut.
The initial idea was to have `decode()` take a `Buf` and `encode()` take a `ButMut`, as they
semantically should. It would have enabled our users to use other buffer backends.

Switching `encode()` to take a BufMut was straightforward, that's done. This solves a gotcha about
using `bytes::BytesMut` vs `bytes::bytes_mut::BytesMut`, and enables the use of other buffer types,
like `Vec`.

Concerning `decode()`, `Buf` currently lacks facilities to rewind the read position
(https://github.com/tokio-rs/bytes/issues/330) and to read a specified number of bytes
(https://github.com/tokio-rs/bytes/issues/129), making it impossible to implement the desired API
unless we assume that our user is using a continuous-memory backend like BytesMut (and even then,
the implementation isn't really pretty).

It is possible to have `decode()` take a `Bytes` (as it semantically should) instead of a
`BytesMut`, but because users typically want to pass a `BytesMut` to `TcpStrem.read()` it would
require a `.freeze()` before passing to `decode()`, which is cumbersome and prevents buffer reuse.

With all that said, this commit does move things in the right direction, and if upstream fixes the
mentioned bugs it'll be that much easyer to switch.